- move in to Zend/Mvc/Service - patch some coding standards - rename config key from 'db' to 'db_adapter_manager'
